RabbitMQ (checkout to Notification)
