--- notification-service/src/main/java/com/example/demo/core/domain/model/KindOfNotification.java
@@ -1,5 +1,0 @@
-package com.example.demo.core.domain.model;
-
-public enum KindOfNotification {
-    WELCOME, ORDER;
-}
