--- notification-service/src/main/java/com/example/demo/core/domain/service/impl/SMSService.java
@@ -1,4 +1,0 @@
-package com.example.demo.core.domain.service.impl;
-
-public class SMSService {
-}
