--- notification-service/src/main/java/com/example/demo/core/domain/service/interfaces/INotificationRepository.java
@@ -1,12 +1,0 @@
-package com.example.demo.core.domain.service.interfaces;
-
-import com.example.demo.core.domain.model.Notification;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.UUID;
-
-@Repository
-public interface INotificationRepository extends JpaRepository<Notification, UUID> {}
-
-
