--- notification-service/src/main/java/com/example/demo/port/user/exception/NotificationNotFoundException.java
@@ -1,10 +1,0 @@
-package com.example.demo.port.user.exception;
-
-import java.util.UUID;
-
-public class NotificationNotFoundException extends RuntimeException {
-    public NotificationNotFoundException(UUID id) {
-        super("Could not find notification " +id);
-    }
-}
-
