+++ notification-service/src/main/java/com/example/notification/core/domain/model/KindOfNotification.java
@@ -1,0 +1,5 @@
+package com.example.notification.core.domain.model;
+
+public enum KindOfNotification {
+    WELCOME, ORDER;
+}
