+++ notification-service/src/main/java/com/example/notification/core/domain/service/impl/EmailService.java
@@ -1,0 +1,56 @@
+package com.example.notification.core.domain.service.impl;
+
+import com.example.notification.core.domain.service.interfaces.IEmailService;
+import jakarta.mail.MessagingException;
+import jakarta.mail.internet.MimeMessage;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.core.io.FileSystemResource;
+import org.springframework.mail.SimpleMailMessage;
+import org.springframework.mail.javamail.JavaMailSender;
+import org.springframework.mail.javamail.MimeMessageHelper;
+import org.springframework.stereotype.Component;
+
+import java.io.File;
+
+@Component
+public class EmailService implements IEmailService {
+
+    @Autowired
+    private JavaMailSender emailSender;
+    private final String EmailAddress = "SA.ComicShop@gmail.com";
+
+    @Override
+    public void sendEmail(
+            String to, String subject, String text) {
+        System.out.println("IN EMAIL SEND METHOD");
+
+        SimpleMailMessage message = new SimpleMailMessage();
+        message.setFrom(EmailAddress);
+        message.setTo(to);
+        message.setSubject(subject);
+        message.setText(text);
+        emailSender.send(message);
+
+    }
+
+    @Override
+    public void sendEmail(
+            String to, String subject, String text, String pathToAttachment) throws MessagingException {
+
+        MimeMessage message = emailSender.createMimeMessage();
+
+        MimeMessageHelper helper = new MimeMessageHelper(message, true);
+
+        helper.setFrom(EmailAddress);
+        helper.setTo(to);
+        helper.setSubject(subject);
+        helper.setText(text);
+
+        FileSystemResource file
+                = new FileSystemResource(new File(pathToAttachment));
+        helper.addAttachment("Invoice", file);
+
+        emailSender.send(message);
+    }
+}
+
