+++ notification-service/src/main/java/com/example/notification/core/domain/service/impl/SMSService.java
@@ -1,0 +1,4 @@
+package com.example.notification.core.domain.service.impl;
+
+public class SMSService {
+}
