+++ notification-service/src/main/java/com/example/notification/port/user/exception/NotificationNotFoundException.java
@@ -1,0 +1,10 @@
+package com.example.notification.port.user.exception;
+
+import java.util.UUID;
+
+public class NotificationNotFoundException extends RuntimeException {
+    public NotificationNotFoundException(UUID id) {
+        super("Could not find notification " +id);
+    }
+}
+
