+++ notification-service/src/test/java/com/example/notification/DemoApplicationTests.java
@@ -1,0 +1,30 @@
+package com.example.notification;
+
+//@SpringBootTest
+class DemoApplicationTests {
+
+/*
+	@Test
+	void contextLoads() {
+	}
+
+    private  INotificationService notificationService;
+
+    @Test
+    void sendMessageTest(){
+
+        UUID userID = UUID.randomUUID();
+        String download = "click here to download";
+        UUID orderID = UUID.randomUUID();
+        KindOfNotification welcomeKind = KindOfNotification.WELCOME;
+        KindOfNotification orderKind = KindOfNotification.ORDER;
+
+        Notification welcomeNotification = new Notification(userID, welcomeKind);
+        Notification orderNotification = new Notification(userID, orderID, download, orderKind);
+
+        notificationService.createNotification(welcomeNotification);
+        notificationService.createNotification(orderNotification);
+    }
+
+ */
+}
